TextMe/requirements.txt
anthropic>=0.40.0
flask>=3.0.0
twilio>=9.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
